django_server/requirements.txt
django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
requests>=2.31.0
nanoid>=2.0.0
